Refactored phoneNumber from Int to Long (int != nullable), added more user integration test, working on authenticate endpoint integration tests.

+++ src/main/java/nl/yourivb/TicketTrack/dtos/AppUser/AppUserPatchDto.java
@@ -7,7 +7,7 @@ public class AppUserPatchDto {
     @Size(min = 2, max = 255)
     private String name;
 
-    private int phoneNumber;
+    private Long phoneNumber;
 
     @Email
     @Size(min = 2, max = 255)
@@ -27,11 +27,11 @@ public class AppUserPatchDto {
         this.name = name;
     }
 
-    public int getPhoneNumber() {
+    public Long getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(Long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
